+++ exonum/src/helpers/metrics.rs
@@ -47,7 +47,6 @@ macro_rules! metric {
 #[doc(hidden)]
 #[allow(unused_variables)]
 pub fn add_metric(metric_name: &str, value: i64) {
-    // TODO: Convert time to some meaningful format.
     let time = format!("{:?}", Utc::now());
 
     #[cfg(feature = "metrics-log")]